+++ embulk-core/src/main/java/org/embulk/EmbulkEmbed.java
@@ -16,11 +16,11 @@ import org.embulk.config.ConfigDiff;
 import org.embulk.config.ConfigLoader;
 import org.embulk.config.ConfigSource;
 import org.embulk.config.DataSource;
-import org.embulk.deps.buffer.PooledBufferAllocator;
 import org.embulk.exec.BulkLoader;
 import org.embulk.exec.ExecutionResult;
 import org.embulk.exec.GuessExecutor;
 import org.embulk.exec.PartialExecutionException;
+import org.embulk.exec.PooledBufferAllocator;
 import org.embulk.exec.PreviewExecutor;
 import org.embulk.exec.PreviewResult;
 import org.embulk.exec.ResumeState;
